requirements.txt
arrow==1.2.3
bcrypt==4.0.1
canvasapi==3.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
Flask==2.2.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
paramiko==3.0.0
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.12.0
